基æ临时提交完成网站搭建

+++ src/main/java/com/hx199513/community/community/model/Question.java
@@ -7,14 +7,14 @@ import lombok.Data;
  */
 @Data
 public class Question {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creator;
-    private Integer viewCount;
-    private Integer commentCount;
-    private Integer likeCount;
+    private Long creator;
+    private Long viewCount;
+    private Long commentCount;
+    private Long likeCount;
 }
